+++ client/config/index.js
@@ -26,6 +26,10 @@ module.exports = {
         target: 'http://localhost:3000',
         changeOrigin: true,
       },
+      '/cable': {
+        target: 'http://localhost:3000',
+        ws: true,
+      },
     },
     // CSS Sourcemaps off by default because relative paths are "buggy"
     // with this option, according to the CSS-Loader README