+++ CHANGELOG.md
@@ -8,12 +8,17 @@
 - Task's state is synced with its related project on `PATCH /api/tasks/:id`
   (please refer to tasks API documentation to know more about it)
 - Remove `project.isInProgress` (equivalent to `project.state === 'started'`)
+- Provide `GET /api/projects/:id` endpoint
+- Provide `GET /api/tasks/:id` endpoint
+- `project.createdAt` and `project.updatedAt` are added
+- `task.createdAt` and `task.updatedAt` are added
 
 ### Migration notes
 
 - `Dockerfile.prod` is renamed in `Dockerfile`
 - Remove unofficial support of Ruby 2.3 (for test suite)
 - Repository moves from https://github.com/marienfressinaud/lessy to https://github.com/lessy-community/lessy
+- Redis is now required, please read specific section in production documentation to know more about it
 
 ## Aquarius 2017-12-29